backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.4
python-multipart==0.0.6
pymupdf==1.23.26
python-docx==1.1.0
langchain==0.1.0
langchain-text-splitters==0.0.1
sentence-transformers==2.2.2